Reorganize texture_atlas module
* Split atlas types into separate files
* Add stubs for array and bindless atlas
* Docs: Ignore experimental modules and functions
* Docs: Fix build issues
