+++ util/update_quick_index.py
@@ -51,6 +51,9 @@ titles = {
     'texture/spritesheet.py': ['Texture Management', 'texture.rst'],
     'texture/tools.py': ['Texture Management', 'texture.rst'],
     'texture/transforms.py': ['Texture Transforms', 'texture_transforms.rst'],
+    'texture_atlas/__init__.py': ['Texture Atlas', 'texture_atlas.rst'],
+    'texture_atlas/base.py': ['Texture Atlas', 'texture_atlas.rst'],
+    'texture_atlas/atlas_2d.py': ['Texture Atlas', 'texture_atlas.rst'],
     'math.py': ['Math', 'math.rst'],
     'types.py': ['Types', 'types.rst'],
     'easing.py': ['Easing', 'easing.rst'],
@@ -62,9 +65,6 @@ titles = {
     'utils.py': ['Misc Utility Functions', 'utility.rst'],
     'window_commands.py': ['Window and View', 'window.rst'],
     'sections.py': ['Window and View', 'window.rst'],
-    'texture_atlas/__init__.py': ['Texture Atlas', 'texture_atlas.rst'],
-    'texture_atlas/base.py': ['Texture Atlas', 'texture_atlas.rst'],
-    'texture_atlas/helpers.py': ['Texture Atlas', 'texture_atlas.rst'],
     'scene.py': ['Sprite Scenes', 'sprite_scenes.rst'],
 
     'tilemap/tilemap.py': ['Tiled Map Reader', 'tilemap.rst'],
@@ -104,15 +104,19 @@ titles = {
     'gl/vertex_array.py': ['OpenGL Vertex Array (VAO)', 'open_gl.rst'],
 }
 excluded_modules = [
-    'version.py'
+    'version.py',
+    'texture_atlas/atlas_array.py',
+    'texture_atlas/atlas_bindless.py',
+    'texture_atlas/helpers.py',
 ]
 
 # Module and class members to exclude
 EXCLUDED_MEMBERS = [
     "ImageData",
-    "AtlasRegion",
-    "ImageDataRefCounter",
     "FakeImage",
+    "load_atlas",
+    "save_atlas",
+    "ImageDataRefCounter",
 ]
 
 def get_member_list(filepath):
@@ -190,7 +194,7 @@ def process_directory(directory: Path, quick_index_file):
             "arcade": "arcade",
             "sprite": "arcade",
             "texture": "arcade",
-            "texture_atlas": "arcade",
+            "texture_atlas": "arcade.texture_atlas",
             "sprite_list": "arcade",
             "text": "arcade",
             "gui": "arcade.gui",
@@ -211,6 +215,7 @@ def process_directory(directory: Path, quick_index_file):
         package = mapping.get(path.name, None) or mapping.get(directory.name, None)
 
         path_name = prepend + path.name
+        # print(package, path.name, path_name)
 
         if path_name in titles and (len(type_list) > 0 or len(class_list) > 0 or len(function_list) > 0):
             title = titles[path_name][0]
